Fix hardware signs
Some hardware was inverted with a `.setInverted(true)` call, to then
have whatever it was set to negated to make it go the right direction. I

There should be a defined polarity to the hardware (+voltage is shooting
on shooter, -voltage is unjamming), (+voltage is up on elevator,
-voltage is down). The polarity is something you get to set the standard
on.

For some cases its relatively arbitary, but when there is a far more
common direction (like shooting and intaking) or a coordinate system
that fits the problem nicely (like elevators or for an angular system,
an arm), its probably best to chose a direction that makes sense to most
people.